Remove the pre-Java 7 code related to the file attributes
The classes related to file attributes are depending on
features introduced in Java 7. To support Java version prior to 7,
plexus IO uses reflection check to choose between two
different implementations. The legacy one that parses the output
from the `ls` command and the Java 7 one that is pure Java.

Because now Java 7 is the minimum required version, remove all the
reflection Java version checks and remove any legacy code that
executes only for Java versions below 7.

There were two version of `FileAttributes` - `FileAttributes` and
`Java7FileAttributes`. Remove the legacy `FileAttributes` and rename
`Java7FileAttributes` to `FileAttributes`.
Rename `Java7AttributeUtils` to `AttributeUtils` as there is
no longer Java version specific implementations for
attributes related classes.

NOTE: The commit diff shows `FileAttributes` as changed
      but it's actually deleted and `Java7FileAttributes`
      is renamed to `FileAttributes`.
      The same goes for `FileAttributesTest` and `Java7FileAttributesTest`

Also remove `PlexusIoResourceAttributeUtils#getFileAttributesByPath( File dir, boolean recursive, boolean includeNumericUserId )`.
`includeNumericUserId` is related to the `ls` command and is no longer needed.

Update the tests. Delete no longer needed resources and tests.
All tests and resources related to the `ls` output parsing
are no longer needed. Also replace
`PlexusIoResourceAttributeUtilsTest#testPermissionDenied`
with `testNonExistingFolder` - when folder is missing
we no longer get `cannot access` response from
the `ls` command and the new test method name is more suitable.
