+++ src/test/java/org/codehaus/plexus/components/io/attributes/PlexusIoResourceAttributeUtilsTest.java
@@ -16,23 +16,15 @@ package org.codehaus.plexus.components.io.attributes;
  * limitations under the License.
  */
 
-import junit.framework.Assert;
-import org.codehaus.plexus.components.io.attributes.AttributeParser.NumericUserIDAttributeParser;
 import org.codehaus.plexus.util.Os;
 import org.codehaus.plexus.util.StringUtils;
 import org.codehaus.plexus.util.cli.CommandLineException;
 import org.codehaus.plexus.util.cli.CommandLineUtils;
 import org.codehaus.plexus.util.cli.Commandline;
-import org.codehaus.plexus.util.cli.StreamConsumer;
 
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.net.URL;
-import java.util.Locale;
 import java.util.Map;
 
 import junit.framework.TestCase;
@@ -43,25 +35,6 @@ public class PlexusIoResourceAttributeUtilsTest
     extends TestCase
 {
 
-    private Locale origSystemLocale;
-
-    @Override
-    protected void setUp()
-        throws Exception
-    {
-       this.origSystemLocale = Locale.getDefault();
-       // sample ls output files have US date format and we use SimpleDateFormt with system locale for ls date format parsing 
-       // otherwise test could fail on systems with non-US locales
-       Locale.setDefault( Locale.US );
-    }
-    
-    @Override
-    protected void tearDown()
-        throws Exception
-    {
-        Locale.setDefault( origSystemLocale );
-    }
-    
     public void testGetAttributesForThisTestClass()
         throws IOException
     {
@@ -82,10 +55,10 @@ public class PlexusIoResourceAttributeUtilsTest
 
         File f = new File( resource.getPath().replaceAll( "%20", " " ) );
 
-        Map attrs =
-            PlexusIoResourceAttributeUtils.getFileAttributesByPath( f, true, true );
+        Map<String, PlexusIoResourceAttributes> attrs =
+            PlexusIoResourceAttributeUtils.getFileAttributesByPath( f, true );
 
-        PlexusIoResourceAttributes fileAttrs = (PlexusIoResourceAttributes) attrs.get( f.getAbsolutePath() );
+        PlexusIoResourceAttributes fileAttrs = attrs.get( f.getAbsolutePath() );
 
         System.out.println( "Got attributes for: " + f.getAbsolutePath() + fileAttrs );
 
@@ -94,7 +67,7 @@ public class PlexusIoResourceAttributeUtilsTest
         assertEquals( System.getProperty( "user.name" ), fileAttrs.getUserName() );
     }
 
-    public void testFolderJava7()
+    public void testFolder()
             throws IOException, CommandLineException {
 
         if (Os.isFamily( Os.FAMILY_WINDOWS ) || Os.isFamily( Os.FAMILY_WIN9X )){
@@ -117,9 +90,10 @@ public class PlexusIoResourceAttributeUtilsTest
         commandLine.addArguments(new String[]{"763", f.getAbsolutePath()});
 
         CommandLineUtils.executeCommandLine(commandLine, null , null);
-        Map attrs = PlexusIoResourceAttributeUtils.getFileAttributesByPath( aDir, true, true );
+        Map<String, PlexusIoResourceAttributes> attrs =
+            PlexusIoResourceAttributeUtils.getFileAttributesByPath( aDir, true );
 
-        PlexusIoResourceAttributes fileAttrs = (PlexusIoResourceAttributes) attrs.get( f.getAbsolutePath() );
+        PlexusIoResourceAttributes fileAttrs = attrs.get( f.getAbsolutePath() );
 
         assertTrue( fileAttrs.isGroupReadable());
         assertTrue( fileAttrs.isGroupWritable());
@@ -136,60 +110,16 @@ public class PlexusIoResourceAttributeUtilsTest
         assertNotNull(fileAttrs);
     }
 
-
-    public void testAttributeParsers()
-    {
-        assertTrue( PlexusIoResourceAttributeUtils.totalLinePattern.matcher( "totalt 420" ).matches() );
-        assertTrue( PlexusIoResourceAttributeUtils.totalLinePattern.matcher( "total 420" ).matches() );
-        assertTrue( PlexusIoResourceAttributeUtils.totalLinePattern.matcher( "JSHS 420" ).matches() );
-    }
-
-    // to make a new testcase
-    // Checkout plexus-io, and from the root of the module type the following:
-    // ls -1nlaR src/main/java/org/codehaus/plexus/components >src/test/resources/`uname`-p1.txt
-    // ls -1laR src/main/java/org/codehaus/plexus/components >src/test/resources/`uname`-p2.txt
-    // Then a test-method that tests the output-
-
-    public void ignoredTestParserUbuntu10_04_en()
-        throws Exception
-    {
-        final Map map = checkStream( "Linux" );
-
-        final FileAttributes o = (FileAttributes) map.get(
-            "src/main/java/org/codehaus/plexus/components/io/attributes/AttributeConstants.java" );
-
-        // -rw-r--r--  1 1020 1030   11108 Mar 16 22:42 build.xml
-        assertEquals( "-rw-rw-r--", new String( o.getLsModeParts() ) );
-        assertEquals( 1020, o.getUserId().intValue() );
-        assertEquals( 1030, o.getGroupId().intValue() );
-        // Should probably test pass 2 too...
-    }
-
-    public void testANotherOneWithTrouble()
-        throws Exception
-    {
-        String output = "drwxr-xr-x+ 13 kristian  staff  442 Oct 11 14:14 ..";
-        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream( output.getBytes() );
-        AttributeParser parser = getNameBasedParser1la();
-        parse( byteArrayInputStream, parser );
-
-    }
-    public void testSingleLine()
-        throws Exception
-    {
-        String output =
-            "-rw-r--r-- 1 1003 1002 1533 2010-04-23 14:34 /home/bamboo/agent1/xml-data/build-dir/PARALLEL-CH1W/checkout/spi/pom.xml";
-        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream( output.getBytes() );
-        AttributeParser parser = getNumericParser1nla();
-        parse( byteArrayInputStream, parser );
-}
-
     public void testSrcResource()
-        throws IOException
+            throws IOException
     {
+        if (Os.isFamily( Os.FAMILY_WINDOWS ) || Os.isFamily( Os.FAMILY_WIN9X )){
+            return; // Nothing to do here.
+        }
+
         File dir = new File("src/test/resources/symlinks");
         final Map<String, PlexusIoResourceAttributes> fileAttributesByPathScreenScrape =
-            PlexusIoResourceAttributeUtils.getFileAttributesByPathScreenScrape( dir, true, true );
+                PlexusIoResourceAttributeUtils.getFileAttributesByPath( dir, true );
         assertNotNull( fileAttributesByPathScreenScrape);
         PlexusIoResourceAttributes pr = null;
         for ( String s : fileAttributesByPathScreenScrape.keySet() )
@@ -200,117 +130,16 @@ public class PlexusIoResourceAttributeUtilsTest
 
         assertTrue(pr.getOctalMode() > 0);
     }
-    public void testPermissionDenied()
+    public void testNonExistingDirectory()
     {
-        File dir = new File("src/test/lsPerms");
+        File dir = new File("src/test/noSuchDirectory");
         try {
-            PlexusIoResourceAttributeUtils.getFileAttributesByPathScreenScrape( dir, true,
-                                                                                true );
-            // We could search for permission denied here and throw some kind of specialized exception.
+            PlexusIoResourceAttributeUtils.getFileAttributesByPath( dir, true );
             fail("We were supposed to get an io exceptions");
         } catch (IOException ignore)
         {
             ignore.printStackTrace();
         }
-    }
-
-    public void testReversedMonthDayOrder()
-        throws Exception
-    {
-        String output = //
-            "-rw-r--r--   1 501  80  7683 31 May 10:06 pom_newer.xml\n" + //
-            "-rwxr--r--   1 502  81  7683  1 Jun 2010  pom_older.xml";
-        InputStream byteArrayInputStream = new ByteArrayInputStream( output.getBytes() );
-        NumericUserIDAttributeParser parser = getNumericParser1nla();
-        parse( byteArrayInputStream, parser );
-        Map<String, PlexusIoResourceAttributes> map = parser.getAttributesByPath();
-        
-        // 6 months or newer ls date format
-        FileAttributes newerFileAttr = (FileAttributes) map.get( "pom_newer.xml" );
-        assertNotNull( newerFileAttr );
-        assertEquals( "-rw-r--r--", new String( newerFileAttr.getLsModeParts() ) );
-        assertEquals( 501, newerFileAttr.getUserId().intValue() );
-        assertEquals( 80, newerFileAttr.getGroupId().intValue() );
-        
-        // older than 6 months ls date format
-        FileAttributes olderFileAttr = (FileAttributes) map.get( "pom_older.xml" );
-        assertNotNull( olderFileAttr );
-        assertEquals( "-rwxr--r--", new String( olderFileAttr.getLsModeParts() ) );
-        assertEquals( 502, olderFileAttr.getUserId().intValue() );
-        assertEquals( 81, olderFileAttr.getGroupId().intValue() );
-    }
-
-    public void testOddLinuxFormatWithExtermelyLargeNumericsSingleLine()
-        throws Exception
-    {
-        String output =
-            "-rw-rw-r-- 1 4294967294 4294967294 7901 2011-06-07 18:39 /mnt/work/src/maven-plugins-trunk/maven-compiler-plugin/pom.xml";
-        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream( output.getBytes() );
-        AttributeParser parser = getNumericParser1nla();
-        parse( byteArrayInputStream, parser );
-    }
-
-    public void testMacOsWithSpace() throws Exception {
-        String output =
-                "-rwxrw--wx  1 kristian  DOMZZZ\\Domain Users  15387 Oct 20 15:30 /Users/kristian/lsrc/plexus/plexus-io/target/test-classes/org/codehaus/plexus/components/io/attributes/PlexusIoResourceAttributeUtilsTest.class";
-        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream( output.getBytes() );
-        AttributeParser parser = getNameBasedParser1la();
-        parse( byteArrayInputStream, parser );
-        final Map<String, PlexusIoResourceAttributes> attributesByPath = parser.getAttributesByPath();
-        String path = "/Users/kristian/lsrc/plexus/plexus-io/target/test-classes/org/codehaus/plexus/components/io/attributes/";
-    }
-
-    public void testMacOsLine()
-        throws Exception
-    {
-        String line = "drwxr-xr-x+ 3 1000  20  102 Oct 11 14:24 ..";
-        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream( line.getBytes() );
-        AttributeParser parser = getNumericParser1nla();
-        parse( byteArrayInputStream, parser );
-
-    }
-    public void ignoredTestParserCygwin()
-        throws Exception
-    {
-        final Map map = checkStream( "CYGWIN_NT-5.1" );
-        final FileAttributes o = (FileAttributes) map.get(
-            "src/main/java/org/codehaus/plexus/components/io/attributes/AttributeConstants.java" );
-
-        // -rw-r--r--  1 1020 1030   11108 Mar 16 22:42 build.xml
-        assertEquals( "-rw-r--r--", new String( o.getLsModeParts() ) );
-        assertEquals( 203222, o.getUserId().intValue() );
-        assertEquals( 10513, o.getGroupId().intValue() );
-    }
-
-    public void testParserSolaris()
-        throws Exception
-    {
-        checkStream( "SunOS" );
-    }
-
-    public void testMisc()
-        throws Exception
-    {
-        checkStream( "Test" );
-    }
-
-    public void testParserFreeBsd()
-        throws Exception
-    {
-        checkStream( "FreeBSD" );
-    }
-
-    public void testParserb26()
-        throws Exception
-    {
-        final Map<String, PlexusIoResourceAttributes> map = checkStream2( "b26-linux" );
-        PlexusIoResourceAttributes attrs = null;
-        for ( String s : map.keySet() )
-        {
-            if (s.endsWith( "App.java" )) attrs = map.get(s);
-        }
-        Assert.assertEquals( 0664, attrs.getOctalMode() );
-
     }
 
     public void testMergeAttributesWithNullBase()
@@ -528,86 +357,4 @@ public class PlexusIoResourceAttributeUtilsTest
         assertEquals( 0, attributes.getOctalMode() );
     }
 
-    private InputStream getStream( String s )
-    {
-        return this.getClass().getClassLoader().getResourceAsStream( s );
-    }
-
-    private Map checkStream( String baseName )
-        throws Exception
-    {
-
-
-        AttributeParser.NumericUserIDAttributeParser numericParser = getNumericParser1nla();
-        InputStream phase1 = getStream( baseName + "-p1.txt" );
-        parse( phase1, numericParser );
-
-
-        final AttributeParser.SymbolicUserIDAttributeParser nameBasedParser = getNameBasedParser1la();
-        InputStream phase2 = getStream( baseName + "-p2.txt" );
-        parse( phase2, nameBasedParser );
-
-        return nameBasedParser.merge( numericParser );
-    }
-
-    private Map<String, PlexusIoResourceAttributes> checkStream2( String baseName )
-        throws Exception
-    {
-
-
-        AttributeParser.NumericUserIDAttributeParser numericParser = getNumericParser1nla();
-        InputStream phase1 = getStream( baseName + "-1nlaR.txt" );
-        parse( phase1, numericParser );
-
-
-        final AttributeParser.SymbolicUserIDAttributeParser nameBasedParser = getNameBasedParser1la();
-        InputStream phase2 = getStream( baseName + "-1laR.txt" );
-        parse( phase2, nameBasedParser );
-
-        return nameBasedParser.merge( numericParser );
-    }
-
-    private AttributeParser.NumericUserIDAttributeParser getNumericParser1nla()
-    {
-        return new AttributeParser.NumericUserIDAttributeParser( createConsoleLogger() );
-    }
-
-    private AttributeParser.SymbolicUserIDAttributeParser getNameBasedParser1la()
-    {
-        return new AttributeParser.SymbolicUserIDAttributeParser( createConsoleLogger() );
-    }
-
-    private StreamConsumer createConsoleLogger()
-    {
-        return new StreamConsumer()
-        {
-            public void consumeLine( String line )
-            {
-                System.out.println(line);
-            }
-        };
-    }
-
-    private void parse( InputStream stream, AttributeParser parser )
-        throws Exception
-    {
-        BufferedReader bufferedReader = new BufferedReader( new InputStreamReader( stream ) );
-        String line = bufferedReader.readLine();
-        int lineNum = 0;
-        try
-        {
-            while ( line != null )
-            {
-                parser.consumeLine( line );
-                line = bufferedReader.readLine();
-                lineNum++;
-            }
-        }
-        catch ( Exception e )
-        {
-            Exception exception = new Exception( "At line " + lineNum + "in source:" + line );
-            exception.initCause( e );
-            throw e;
-        }
-    }
 }
