+++ src/test/java/org/codehaus/plexus/components/io/resources/PlexusIoPlexusIoFileResourceTest.java
@@ -3,8 +3,7 @@ package org.codehaus.plexus.components.io.resources;
 import java.io.File;
 import java.io.IOException;
 
-import org.codehaus.plexus.components.io.attributes.Java7FileAttributes;
-import org.codehaus.plexus.components.io.attributes.Java7Reflector;
+import org.codehaus.plexus.components.io.attributes.FileAttributes;
 import org.codehaus.plexus.components.io.attributes.PlexusIoResourceAttributes;
 
 import junit.framework.TestCase;
@@ -17,10 +16,9 @@ public class PlexusIoPlexusIoFileResourceTest
     public void testRealSymlink()
         throws IOException
     {
-        if (!Java7Reflector.isAtLeastJava7()) return;
         if ( Os.isFamily(Os.FAMILY_WINDOWS)) return;
         final File file = new File( "src/test/resources/symlinks/src/symDir" );
-        PlexusIoResourceAttributes attrs = Java7FileAttributes.uncached( file );
+        PlexusIoResourceAttributes attrs = FileAttributes.uncached( file );
         assertTrue(attrs.isSymbolicLink());
         PlexusIoFileResource r = new PlexusIoFileResource( file,  "symDir", attrs);
         assertTrue(r.isSymbolicLink());
