--- src/test/resources/CYGWIN_NT-5.1-p2.txt
@@ -1,77 +1,0 @@
-src/main/java/org/codehaus/plexus/components:
-total 0
-drwxr-xr-x+ 3 x575722 Domain Users 0 Apr 21 08:49 .
-drwxr-xr-x+ 3 x575722 Domain Users 0 Apr 21 08:49 ..
-drwxr-xr-x+ 6 x575722 Domain Users 0 Apr 21 08:49 io
-
-src/main/java/org/codehaus/plexus/components/io:
-total 0
-drwxr-xr-x+ 6 x575722 Domain Users 0 Apr 21 08:49 .
-drwxr-xr-x+ 3 x575722 Domain Users 0 Apr 21 08:49 ..
-drwxr-xr-x+ 2 x575722 Domain Users 0 Apr 23 08:24 attributes
-drwxr-xr-x+ 2 x575722 Domain Users 0 Apr 21 08:49 filemappers
-drwxr-xr-x+ 2 x575722 Domain Users 0 Apr 21 08:49 fileselectors
-drwxr-xr-x+ 3 x575722 Domain Users 0 Apr 21 08:49 resources
-
-src/main/java/org/codehaus/plexus/components/io/attributes:
-total 48
-drwxr-xr-x+ 2 x575722 Domain Users     0 Apr 23 08:24 .
-drwxr-xr-x+ 6 x575722 Domain Users     0 Apr 21 08:49 ..
--rw-r--r--  1 x575722 Domain Users  1242 Apr 21 08:49 AttributeConstants.java
--rw-r--r--  1 x575722 Domain Users  9565 Apr 21 08:49 FileAttributes.java
--rw-r--r--  1 x575722 Domain Users 17872 Apr 23 08:24 PlexusIoResourceAttributeUtils.java
--rw-r--r--  1 x575722 Domain Users  2179 Apr 21 08:49 PlexusIoResourceAttributes.java
--rw-r--r--  1 x575722 Domain Users  5266 Apr 21 08:49 SimpleResourceAttributes.java
-
-src/main/java/org/codehaus/plexus/components/io/filemappers:
-total 36
-drwxr-xr-x+ 2 x575722 Domain Users    0 Apr 21 08:49 .
-drwxr-xr-x+ 6 x575722 Domain Users    0 Apr 21 08:49 ..
--rw-r--r--  1 x575722 Domain Users 1293 Apr 21 08:49 AbstractFileMapper.java
--rw-r--r--  1 x575722 Domain Users 2516 Apr 21 08:49 FileExtensionMapper.java
--rw-r--r--  1 x575722 Domain Users 1358 Apr 21 08:49 FileMapper.java
--rw-r--r--  1 x575722 Domain Users 1384 Apr 21 08:49 FlattenFileMapper.java
--rw-r--r--  1 x575722 Domain Users 1216 Apr 21 08:49 IdentityMapper.java
--rw-r--r--  1 x575722 Domain Users 2013 Apr 21 08:49 MergeFileMapper.java
--rw-r--r--  1 x575722 Domain Users 1665 Apr 21 08:49 PrefixFileMapper.java
--rw-r--r--  1 x575722 Domain Users 2497 Apr 21 08:49 RegExpFileMapper.java
--rw-r--r--  1 x575722 Domain Users  877 Apr 21 08:49 package.html
-
-src/main/java/org/codehaus/plexus/components/io/fileselectors:
-total 24
-drwxr-xr-x+ 2 x575722 Domain Users    0 Apr 21 08:49 .
-drwxr-xr-x+ 6 x575722 Domain Users    0 Apr 21 08:49 ..
--rw-r--r--  1 x575722 Domain Users 1035 Apr 21 08:49 AllFilesFileSelector.java
--rw-r--r--  1 x575722 Domain Users 1761 Apr 21 08:49 FileInfo.java
--rw-r--r--  1 x575722 Domain Users 1377 Apr 21 08:49 FileSelector.java
--rw-r--r--  1 x575722 Domain Users 8145 Apr 21 08:49 IncludeExcludeFileSelector.java
--rw-r--r--  1 x575722 Domain Users  616 Apr 21 08:49 package.html
-
-src/main/java/org/codehaus/plexus/components/io/resources:
-total 65
-drwxr-xr-x+ 3 x575722 Domain Users    0 Apr 21 08:49 .
-drwxr-xr-x+ 6 x575722 Domain Users    0 Apr 21 08:49 ..
--rw-r--r--  1 x575722 Domain Users 2277 Apr 21 08:49 AbstractPlexusIoArchiveResourceCollection.java
--rw-r--r--  1 x575722 Domain Users 2884 Apr 21 08:49 AbstractPlexusIoResource.java
--rw-r--r--  1 x575722 Domain Users 6776 Apr 21 08:49 AbstractPlexusIoResourceCollection.java
--rw-r--r--  1 x575722 Domain Users 1243 Apr 21 08:49 AbstractPlexusIoResourceWithAttributes.java
--rw-r--r--  1 x575722 Domain Users 1542 Apr 21 08:49 PlexusIOResourceCollectionWithAttributes.java
--rw-r--r--  1 x575722 Domain Users 1042 Apr 21 08:49 PlexusIoArchivedResourceCollection.java
--rw-r--r--  1 x575722 Domain Users 2980 Apr 21 08:49 PlexusIoCompressedFileResourceCollection.java
--rw-r--r--  1 x575722 Domain Users 3355 Apr 21 08:49 PlexusIoFileResource.java
--rw-r--r--  1 x575722 Domain Users 6054 Apr 21 08:49 PlexusIoFileResourceCollection.java
--rw-r--r--  1 x575722 Domain Users 1911 Apr 21 08:49 PlexusIoResource.java
--rw-r--r--  1 x575722 Domain Users 2127 Apr 21 08:49 PlexusIoResourceCollection.java
--rw-r--r--  1 x575722 Domain Users  949 Apr 21 08:49 PlexusIoResourceWithAttributes.java
--rw-r--r--  1 x575722 Domain Users 1089 Apr 21 08:49 PlexusIoURLResource.java
--rw-r--r--  1 x575722 Domain Users 2856 Apr 21 08:49 PlexusIoZipFileResourceCollection.java
--rw-r--r--  1 x575722 Domain Users  534 Apr 21 08:49 package.html
-drwxr-xr-x+ 2 x575722 Domain Users    0 Apr 21 08:49 proxy
-
-src/main/java/org/codehaus/plexus/components/io/resources/proxy:
-total 16
-drwxr-xr-x+ 2 x575722 Domain Users    0 Apr 21 08:49 .
-drwxr-xr-x+ 3 x575722 Domain Users    0 Apr 21 08:49 ..
--rw-r--r--  1 x575722 Domain Users 1753 Apr 21 08:49 PlexusIoProxyResource.java
--rw-r--r--  1 x575722 Domain Users 6701 Apr 21 08:49 PlexusIoProxyResourceCollection.java
--rw-r--r--  1 x575722 Domain Users 1897 Apr 21 08:49 PlexusIoProxyResourceWithAttributes.java
