--- src/test/resources/FreeBSD-p2.txt
@@ -1,76 +1,0 @@
-total 9
-drwxr-xr-x  3 krosenvold  krosenvold  3 Apr 23 06:57 .
-drwxr-xr-x  3 krosenvold  krosenvold  3 Apr 23 06:57 ..
-drwxr-xr-x  6 krosenvold  krosenvold  6 Apr 23 06:57 io
-
-src/main/java/org/codehaus/plexus/components/io:
-total 18
-drwxr-xr-x  6 krosenvold  krosenvold   6 Apr 23 06:57 .
-drwxr-xr-x  3 krosenvold  krosenvold   3 Apr 23 06:57 ..
-drwxr-xr-x  2 krosenvold  krosenvold   7 Apr 23 06:57 attributes
-drwxr-xr-x  2 krosenvold  krosenvold  11 Apr 23 06:57 filemappers
-drwxr-xr-x  2 krosenvold  krosenvold   7 Apr 23 06:57 fileselectors
-drwxr-xr-x  3 krosenvold  krosenvold  18 Apr 23 06:57 resources
-
-src/main/java/org/codehaus/plexus/components/io/attributes:
-total 50
-drwxr-xr-x  2 krosenvold  krosenvold      7 Apr 23 06:57 .
-drwxr-xr-x  6 krosenvold  krosenvold      6 Apr 23 06:57 ..
--rw-r--r--  1 krosenvold  krosenvold   1242 Apr 23 06:57 AttributeConstants.java
--rw-r--r--  1 krosenvold  krosenvold   9565 Apr 23 06:57 FileAttributes.java
--rw-r--r--  1 krosenvold  krosenvold  17872 Apr 23 06:57 PlexusIoResourceAttributeUtils.java
--rw-r--r--  1 krosenvold  krosenvold   2179 Apr 23 06:57 PlexusIoResourceAttributes.java
--rw-r--r--  1 krosenvold  krosenvold   5266 Apr 23 06:57 SimpleResourceAttributes.java
-
-src/main/java/org/codehaus/plexus/components/io/filemappers:
-total 36
-drwxr-xr-x  2 krosenvold  krosenvold    11 Apr 23 06:57 .
-drwxr-xr-x  6 krosenvold  krosenvold     6 Apr 23 06:57 ..
--rw-r--r--  1 krosenvold  krosenvold  1293 Apr 23 06:57 AbstractFileMapper.java
--rw-r--r--  1 krosenvold  krosenvold  2516 Apr 23 06:57 FileExtensionMapper.java
--rw-r--r--  1 krosenvold  krosenvold  1358 Apr 23 06:57 FileMapper.java
--rw-r--r--  1 krosenvold  krosenvold  1384 Apr 23 06:57 FlattenFileMapper.java
--rw-r--r--  1 krosenvold  krosenvold  1216 Apr 23 06:57 IdentityMapper.java
--rw-r--r--  1 krosenvold  krosenvold  2013 Apr 23 06:57 MergeFileMapper.java
--rw-r--r--  1 krosenvold  krosenvold  1665 Apr 23 06:57 PrefixFileMapper.java
--rw-r--r--  1 krosenvold  krosenvold  2497 Apr 23 06:57 RegExpFileMapper.java
--rw-r--r--  1 krosenvold  krosenvold   877 Apr 23 06:57 package.html
-
-src/main/java/org/codehaus/plexus/components/io/fileselectors:
-total 28
-drwxr-xr-x  2 krosenvold  krosenvold     7 Apr 23 06:57 .
-drwxr-xr-x  6 krosenvold  krosenvold     6 Apr 23 06:57 ..
--rw-r--r--  1 krosenvold  krosenvold  1035 Apr 23 06:57 AllFilesFileSelector.java
--rw-r--r--  1 krosenvold  krosenvold  1761 Apr 23 06:57 FileInfo.java
--rw-r--r--  1 krosenvold  krosenvold  1377 Apr 23 06:57 FileSelector.java
--rw-r--r--  1 krosenvold  krosenvold  8145 Apr 23 06:57 IncludeExcludeFileSelector.java
--rw-r--r--  1 krosenvold  krosenvold   616 Apr 23 06:57 package.html
-
-src/main/java/org/codehaus/plexus/components/io/resources:
-total 72
-drwxr-xr-x  3 krosenvold  krosenvold    18 Apr 23 06:57 .
-drwxr-xr-x  6 krosenvold  krosenvold     6 Apr 23 06:57 ..
--rw-r--r--  1 krosenvold  krosenvold  2277 Apr 23 06:57 AbstractPlexusIoArchiveResourceCollection.java
--rw-r--r--  1 krosenvold  krosenvold  2884 Apr 23 06:57 AbstractPlexusIoResource.java
--rw-r--r--  1 krosenvold  krosenvold  6776 Apr 23 06:57 AbstractPlexusIoResourceCollection.java
--rw-r--r--  1 krosenvold  krosenvold  1243 Apr 23 06:57 AbstractPlexusIoResourceWithAttributes.java
--rw-r--r--  1 krosenvold  krosenvold  1542 Apr 23 06:57 PlexusIOResourceCollectionWithAttributes.java
--rw-r--r--  1 krosenvold  krosenvold  1042 Apr 23 06:57 PlexusIoArchivedResourceCollection.java
--rw-r--r--  1 krosenvold  krosenvold  2980 Apr 23 06:57 PlexusIoCompressedFileResourceCollection.java
--rw-r--r--  1 krosenvold  krosenvold  3355 Apr 23 06:57 PlexusIoFileResource.java
--rw-r--r--  1 krosenvold  krosenvold  6054 Apr 23 06:57 PlexusIoFileResourceCollection.java
--rw-r--r--  1 krosenvold  krosenvold  1911 Apr 23 06:57 PlexusIoResource.java
--rw-r--r--  1 krosenvold  krosenvold  2127 Apr 23 06:57 PlexusIoResourceCollection.java
--rw-r--r--  1 krosenvold  krosenvold   949 Apr 23 06:57 PlexusIoResourceWithAttributes.java
--rw-r--r--  1 krosenvold  krosenvold  1089 Apr 23 06:57 PlexusIoURLResource.java
--rw-r--r--  1 krosenvold  krosenvold  2856 Apr 23 06:57 PlexusIoZipFileResourceCollection.java
--rw-r--r--  1 krosenvold  krosenvold   534 Apr 23 06:57 package.html
-drwxr-xr-x  2 krosenvold  krosenvold     5 Apr 23 06:57 proxy
-
-src/main/java/org/codehaus/plexus/components/io/resources/proxy:
-total 20
-drwxr-xr-x  2 krosenvold  krosenvold     5 Apr 23 06:57 .
-drwxr-xr-x  3 krosenvold  krosenvold    18 Apr 23 06:57 ..
--rw-r--r--  1 krosenvold  krosenvold  1753 Apr 23 06:57 PlexusIoProxyResource.java
--rw-r--r--  1 krosenvold  krosenvold  6701 Apr 23 06:57 PlexusIoProxyResourceCollection.java
--rw-r--r--  1 krosenvold  krosenvold  1897 Apr 23 06:57 PlexusIoProxyResourceWithAttributes.java
