--- src/test/resources/SunOS-p1.txt
@@ -1,77 +1,0 @@
-src/main/java/org/codehaus/plexus/components:
-total 6
-drwxrwxr-x   3 10105    10076        512 Apr 23 08:50 .
-drwxrwxr-x   3 10105    10076        512 Apr 23 08:50 ..
-drwxrwxr-x   6 10105    10076        512 Apr 23 08:50 io
-
-src/main/java/org/codehaus/plexus/components/io:
-total 12
-drwxrwxr-x   6 10105    10076        512 Apr 23 08:50 .
-drwxrwxr-x   3 10105    10076        512 Apr 23 08:50 ..
-drwxrwxr-x   2 10105    10076        512 Apr 23 08:50 attributes
-drwxrwxr-x   2 10105    10076        512 Apr 23 08:50 filemappers
-drwxrwxr-x   2 10105    10076        512 Apr 23 08:50 fileselectors
-drwxrwxr-x   3 10105    10076       1024 Apr 23 08:50 resources
-
-src/main/java/org/codehaus/plexus/components/io/attributes:
-total 82
-drwxrwxr-x   2 10105    10076        512 Apr 23 08:50 .
-drwxrwxr-x   6 10105    10076        512 Apr 23 08:50 ..
--rw-rw-r--   1 10105    10076       1242 Apr 23 08:50 AttributeConstants.java
--rw-rw-r--   1 10105    10076       9565 Apr 23 08:50 FileAttributes.java
--rw-rw-r--   1 10105    10076       2179 Apr 23 08:50 PlexusIoResourceAttributes.java
--rw-rw-r--   1 10105    10076      17872 Apr 23 08:50 PlexusIoResourceAttributeUtils.java
--rw-rw-r--   1 10105    10076       5266 Apr 23 08:50 SimpleResourceAttributes.java
-
-src/main/java/org/codehaus/plexus/components/io/filemappers:
-total 42
-drwxrwxr-x   2 10105    10076        512 Apr 23 08:50 .
-drwxrwxr-x   6 10105    10076        512 Apr 23 08:50 ..
--rw-rw-r--   1 10105    10076       1293 Apr 23 08:50 AbstractFileMapper.java
--rw-rw-r--   1 10105    10076       2516 Apr 23 08:50 FileExtensionMapper.java
--rw-rw-r--   1 10105    10076       1358 Apr 23 08:50 FileMapper.java
--rw-rw-r--   1 10105    10076       1384 Apr 23 08:50 FlattenFileMapper.java
--rw-rw-r--   1 10105    10076       1216 Apr 23 08:50 IdentityMapper.java
--rw-rw-r--   1 10105    10076       2013 Apr 23 08:50 MergeFileMapper.java
--rw-rw-r--   1 10105    10076        877 Apr 23 08:50 package.html
--rw-rw-r--   1 10105    10076       1665 Apr 23 08:50 PrefixFileMapper.java
--rw-rw-r--   1 10105    10076       2497 Apr 23 08:50 RegExpFileMapper.java
-
-src/main/java/org/codehaus/plexus/components/io/fileselectors:
-total 34
-drwxrwxr-x   2 10105    10076        512 Apr 23 08:50 .
-drwxrwxr-x   6 10105    10076        512 Apr 23 08:50 ..
--rw-rw-r--   1 10105    10076       1035 Apr 23 08:50 AllFilesFileSelector.java
--rw-rw-r--   1 10105    10076       1761 Apr 23 08:50 FileInfo.java
--rw-rw-r--   1 10105    10076       1377 Apr 23 08:50 FileSelector.java
--rw-rw-r--   1 10105    10076       8145 Apr 23 08:50 IncludeExcludeFileSelector.java
--rw-rw-r--   1 10105    10076        616 Apr 23 08:50 package.html
-
-src/main/java/org/codehaus/plexus/components/io/resources:
-total 94
-drwxrwxr-x   3 10105    10076       1024 Apr 23 08:50 .
-drwxrwxr-x   6 10105    10076        512 Apr 23 08:50 ..
--rw-rw-r--   1 10105    10076       2277 Apr 23 08:50 AbstractPlexusIoArchiveResourceCollection.java
--rw-rw-r--   1 10105    10076       2884 Apr 23 08:50 AbstractPlexusIoResource.java
--rw-rw-r--   1 10105    10076       6776 Apr 23 08:50 AbstractPlexusIoResourceCollection.java
--rw-rw-r--   1 10105    10076       1243 Apr 23 08:50 AbstractPlexusIoResourceWithAttributes.java
--rw-rw-r--   1 10105    10076        534 Apr 23 08:50 package.html
--rw-rw-r--   1 10105    10076       1042 Apr 23 08:50 PlexusIoArchivedResourceCollection.java
--rw-rw-r--   1 10105    10076       2980 Apr 23 08:50 PlexusIoCompressedFileResourceCollection.java
--rw-rw-r--   1 10105    10076       3355 Apr 23 08:50 PlexusIoFileResource.java
--rw-rw-r--   1 10105    10076       6054 Apr 23 08:50 PlexusIoFileResourceCollection.java
--rw-rw-r--   1 10105    10076       1911 Apr 23 08:50 PlexusIoResource.java
--rw-rw-r--   1 10105    10076       2127 Apr 23 08:50 PlexusIoResourceCollection.java
--rw-rw-r--   1 10105    10076       1542 Apr 23 08:50 PlexusIOResourceCollectionWithAttributes.java
--rw-rw-r--   1 10105    10076        949 Apr 23 08:50 PlexusIoResourceWithAttributes.java
--rw-rw-r--   1 10105    10076       1089 Apr 23 08:50 PlexusIoURLResource.java
--rw-rw-r--   1 10105    10076       2856 Apr 23 08:50 PlexusIoZipFileResourceCollection.java
-drwxrwxr-x   2 10105    10076        512 Apr 23 08:50 proxy
-
-src/main/java/org/codehaus/plexus/components/io/resources/proxy:
-total 26
-drwxrwxr-x   2 10105    10076        512 Apr 23 08:50 .
-drwxrwxr-x   3 10105    10076       1024 Apr 23 08:50 ..
--rw-rw-r--   1 10105    10076       1753 Apr 23 08:50 PlexusIoProxyResource.java
--rw-rw-r--   1 10105    10076       6701 Apr 23 08:50 PlexusIoProxyResourceCollection.java
--rw-rw-r--   1 10105    10076       1897 Apr 23 08:50 PlexusIoProxyResourceWithAttributes.java
