--- src/test/resources/b26-linux-1laR.txt
@@ -1,139 +1,0 @@
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule:
-total 140
-drwxrwxr-x  4 release15 release15   4096 Oct 22 19:17 .
-drwxrwxr-x 14 release15 release15   4096 Oct 22 19:14 ..
--rw-rw-r--  1 release15 release15 111326 Oct 22 19:17 build.log
-drwxrwxr-x  4 release15 release15   4096 Oct 22 19:17 child
--rw-rw-r--  1 release15 release15    814 Oct 22 19:14 invoker.properties
--rw-rw-r--  1 release15 release15   1356 Oct 22 19:14 pom.xml
-drwxrwxr-x  3 release15 release15   4096 Oct 22 19:14 src
--rw-rw-r--  1 release15 release15   1289 Oct 22 19:14 verify.bsh
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child:
-total 20
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:17 .
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:17 ..
--rw-rw-r-- 1 release15 release15 1811 Oct 22 19:14 pom.xml
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 src
-drwxrwxr-x 6 release15 release15 4096 Oct 22 19:17 target
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/src:
-total 12
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:17 ..
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:14 main
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/src/main:
-total 16
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 ..
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:14 assembly
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:14 assembly-resources
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/src/main/assembly:
-total 12
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:14 ..
--rw-rw-r-- 1 release15 release15 1389 Oct 22 19:14 src.xml
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/src/main/assembly-resources:
-total 12
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:14 ..
--rw-rw-r-- 1 release15 release15   18 Oct 22 19:14 test.txt
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/target:
-total 28
-drwxrwxr-x 6 release15 release15 4096 Oct 22 19:17 .
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:17 ..
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:17 archive-tmp
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:17 child-1-src
--rw-rw-r-- 1 release15 release15 2206 Oct 22 19:17 child-1.jar
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:17 maven-archiver
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:17 surefire
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/target/archive-tmp:
-total 8
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:17 .
-drwxrwxr-x 6 release15 release15 4096 Oct 22 19:17 ..
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/target/child-1-src:
-total 12
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:17 .
-drwxrwxr-x 6 release15 release15 4096 Oct 22 19:17 ..
----------- 1 release15 release15    1 Oct 22 19:14 test.txt
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/target/maven-archiver:
-total 12
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:17 .
-drwxrwxr-x 6 release15 release15 4096 Oct 22 19:17 ..
--rw-rw-r-- 1 release15 release15  124 Oct 22 19:17 pom.properties
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/child/target/surefire:
-total 20
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:17 .
-drwxrwxr-x 6 release15 release15 4096 Oct 22 19:17 ..
--rw-rw-r-- 1 release15 release15 1449 Oct 22 19:17 surefire6670933677475933695tmp
--rw-rw-r-- 1 release15 release15  623 Oct 22 19:17 surefire8953811750968444590tmp
--rw-rw-r-- 1 release15 release15  849 Oct 22 19:17 surefirebooter6288163577541824457.jar
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src:
-total 12
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:17 ..
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:14 main
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main:
-total 16
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 ..
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:14 assembly
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 java
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main/assembly:
-total 12
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:14 ..
--rw-rw-r-- 1 release15 release15 1482 Oct 22 19:14 bin.xml
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main/java:
-total 12
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 4 release15 release15 4096 Oct 22 19:14 ..
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 org
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main/java/org:
-total 12
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 ..
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 apache
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main/java/org/apache:
-total 12
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 ..
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 maven
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main/java/org/apache/maven:
-total 12
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 ..
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 plugin
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main/java/org/apache/maven/plugin:
-total 12
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 ..
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 assembly
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main/java/org/apache/maven/plugin/assembly:
-total 12
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 ..
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:14 test
-
-/home/release15/maven-plugins/maven-assembly-plugin/target/it/projects/file-sets/filtered-multimodule/src/main/java/org/apache/maven/plugin/assembly/test:
-total 12
-drwxrwxr-x 2 release15 release15 4096 Oct 22 19:14 .
-drwxrwxr-x 3 release15 release15 4096 Oct 22 19:14 ..
--rw-rw-r-- 1 release15 release15 1012 Oct 22 19:14 App.java
